Use floating point for more accurate display of partition sizes

+++ partitionmanager.cpp
@@ -2751,7 +2751,7 @@ void TWPartitionManager::Get_Partition_List(string ListType,
 							Backup_Size += (*subpart)->Backup_Size;
 					}
 				}
-				sprintf(backup_size, "%llu", Backup_Size / 1024 / 1024);
+				sprintf(backup_size, "%.2lf", (double)Backup_Size / 1024 / 1024);
 				part.Display_Name = (*iter)->Backup_Display_Name + " (";
 				part.Display_Name += backup_size;
 				part.Display_Name += "MB)";
